requirements.txt
# Core dependencies
requests==2.31.0
python-dotenv==1.0.0
anthropic==0.18.1

# Video processing
opencv-python-headless==4.8.1.78
moviepy==1.0.3
ffmpeg-python==0.2.0

# AI models
openai-whisper==20231117
ultralytics==8.0.200
mediapipe==0.10.8
easyocr==1.7.0
torch==2.1.0
torchvision==0.16.0
transformers==4.35.2
scenedetect[opencv]==0.6.2

# Object tracking
deep-sort-realtime==1.3.2

# CLIP dependencies
# Note: CLIP itself is installed via: pip install git+https://github.com/openai/CLIP.git
# This is handled automatically by setup/bootstrap.py
ftfy==6.1.3
regex==2023.10.3
tqdm==4.66.1

# Utilities
numpy==1.24.3
pillow==10.1.0
scipy==1.11.4
scikit-learn==1.3.2
